backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Miguel Angelo", email = "miguelrsant01@gmail.com"},
]
dependencies = ["flask>=3.1.2", "gunicorn>=23.0.0", "flask-cors>=6.0.1", "python-dotenv>=1.2.1", "SQLAlchemy>=2.0.44", "Flask-Migrate>=4.1.0", "psycopg2>=2.9.11"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
